Jeremias/codebase/AdversarialLearningProject/ICML/GNNPlus-main/requirements.txt
# Core deep learning
torch
torch_geometric
torch_scatter
torch_sparse
torch_cluster
torch_spline_conv
pytorch_lightning

# Visualization and data handling
pandas
numpy
scikit-learn
matplotlib
seaborn
yacs

# Configuration and experiment management
python-dotenv
wandb # Optional, in case you enable it in the config
ogb